pyproject.toml
[tool.poetry]
name = "nekro-agent"
version = "0.1.13"
description = "Nekro-agent"
authors = ["KroMiose <li_xiangff@163.com>"]
readme = "README.md"
packages = [{ include = "nekro_agent" }]
repository = "https://github.com/KroMiose/nekro-agent"

[tool.poetry.scripts]
publish = "scripts.run_publish:main"
bot = "run_bot:main"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
uvicorn = "^0.23.2"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
nonebot-adapter-onebot = "^2.4.2"
nonebot2 = { extras = ["fastapi"], version = "^2.2.1" }
httpx = "^0.27.0"
openai = ">=0.27.0"
openpyxl = "^3.1.2"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.2.0"
docker = "^7.1.0"

miose-toolkit-common = "^0.3.7"
miose-toolkit-llm = "^0.0.2"
miose-toolkit-logger = "^0.0.7"
miose-toolkit-db = "^0.2.2"
aiodocker = "^0.22.2"

[tool.poetry.group.dev.dependencies]
nb-cli = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = ["nekro_agent"]
plugin_dirs = []
builtin_plugins = []

# [[tool.poetry.source]]
# name = "aliyun"
# url = "https://mirrors.aliyun.com/pypi/simple/"
# priority = "primary"

[tool.black]
line-length = 128

[tool.ruff]
extend-exclude = []
ignore = [
    "E402",
    "E501",
    "E722",
    "B008",
    "B905",
    "F401",
    "TID252",
    "TRY002",
    "TRY003",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF006",
    "RUF012",
    "RUF100",
    "PGH003",
    "N801",
    "A003",
]
select = [
    "F",
    "E",
    "I",
    "B",
    "A",
    "COM",
    "C4",
    "ISC",
    "PIE",
    "PYI",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SIM",
    "TID",
    "ARG",
    "PTH",
    "PGH",
    "TRY",
    "RUF",
]
